requirements.txt
matplotlib==3.10.1
numpy==2.1.1
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
scipy==1.15.2
seaborn==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
scikit-learn==1.6.1
scikit-image==0.25.2
pyyaml==6.0.2
requests==2.32.3
